databse try again

+++ texts/migrations/0008_auto_20151029_1318.py
@@ -14,6 +14,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='feedback',
             name='phone',
-            field=models.IntegerField(verbose_name='Телефoн'),
+            field=models.BigIntegerField(verbose_name='Телефoн'),
         ),
     ]
